refactor(invitation): move BoardInvitationService under Service\Board
Align the service with the App\Service\<Domain> convention established by
the DTO -> Service -> Repository refactor (see docs/REFACTOR_PLAN.md §3).
The class is also made final like the other domain services; nothing mocks
or extends it.
