+++ src/Controller/BoardController.php
@@ -13,7 +13,7 @@ use App\Repository\AccountRepository;
 use App\Repository\BoardRepository;
 use App\Repository\BoardInvitationRepository;
 use App\Service\Board\BoardService;
-use App\Service\BoardInvitationService;
+use App\Service\Board\BoardInvitationService;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
